MGMT-848: Clean up ISOs that are older than a day
Adds a k8s CronJob that lists the objects in the S3 bucket that we use
for storing ISOs and deletes any that are older than 1 day. This is
because Scality does not support bucket lifecycle policies for
expiration. Deployments running in AWS shouldn't use this.

The expirer script has some environment variables that aren't passed
automatically. The defaults are currently suitable for our tests and can
be overidden in other deployments.