MGMT-797: Code changes for storing image properties
We now store a single image per cluster in S3, and store its properties
in the DB so that the user can view them when creating a new image.

We need to avoid a race where we end up having two images created at the
same time and storing the data for one in S3 and the metadata of the
other in the DB. So first of all we wait 10 seconds between requests,
which should be more than enough time for the first request's job to
start. Second, every job kills the job before it if it's still running
before it starts its own job.