Host: Installation Cancel
When canceling a mid-run cluster installation, set the relevant hosts
states to error using a new statemachine transition, only if the are
currently in installing state. Any error will return as an ApiError
object.
The new transition wont override the state info of an host which was
already in error state.
This commit includes unit test of the host cancellation functionality
as well.