MGMT-1099 Subsystem: Add reset cluster tests
The tests will reset a failed installation and try to re-register the
hosts. Second test will test that while cluster is ready or installing
the reset option won't be available. Third will try to reset cluster
that is ready or installing, expected to be failed, forth will try to
reset cluster when hosts in multiple states, expected to get conflict.