Allow registering already installed host
Registering a host which is already installed can be done to the same
cluster or to a different one.

Without this PR, it is already possible to register a host to a different
cluster. However, in this situation there are two hosts with the same
machine-uuid in the DB (hosts.connectivity.remote_hosts.host_id).

This PR enabled the re-registering of the same host to the cluster, in
which case the host status will transit from 'Installed' to
'Discovering' and from 'Discovering' to 'Known'.

Signed-off-by: Moti Asayag <masayag@redhat.com>