Updated use of Postgres/S3 credentials from Secrets.
This change is done in order to be as close as possible to the
integration environment which deploy Postgres and S3 and corresponding
secrets which contain the credentials to connect to them.

Now the credentials for connecting to DB comes from
`assisted-installer/assisted-installer-rds` secret.

The credentials for connecting to S3 comes from
`assisted-installer/assisted-installer-s3` secret.

Also grouped all DB config fields into a separate struct.

Signed-off-by: Yoni Bettan <ybettan@redhat.com>)