Fix tests which rely on execution under three minutes
Previously many of these tests could fail if over three minutes
elapsed between when the array of test cases was defined and when
any individual test was run.

This is because the host `CheckedInAt` time was defined as time.Now()
in the test case definition, not when the test was actually being run.

This was happening in two contexts. One was "All transitions" where
some cases were relying on the check-in time to be out of range (unset)
and others needed it to be within the last 3 minutes. The other was
the "Unique hostname" context where all the tests were using time.Now()
so I removed the test case parameter for that one.