Raz/ocpbugsm 13597 (#50)
* OCPBUGSM-13597 Update register transition

There are some edge cases on installation where user tries
to reset installation on the same time reboot is called. On
some cases  the agent will get to reset itself and register
again just before the reboot and the cluster monitor will not
get to set the status in resetting-pending-user-action on
time. In this commit I added to the host statemachine a
tranistion verifies that host can't register from resetting
state if the stage is reboot.

* OCPBUGSM-13597 Register transition fix

Due ctrl.Finish() added in previews commit the test is now
fails because of an unnecessary event handler call when no
event should be received.