MGMT-1772 Use an env var to set the s3 url scheme if not present (#95)
* Use an env var to set the s3 url scheme if not present

Currently we sometimes are provided an s3 "url" which is just a
hostname. This causes an error when trying to upload the iso file.

This PR will add either http or https to the hostname to create
a valid url.

* Fix the BMConfig S3 endpoint as well as job configs

This will fix 'unsupported protocol scheme ""' errors when running
bminventory.(*bareMetalInventory).GetCredentials