[MGMT-1828] Generate new ISO after proxy was changed
Image should be generated based on its recent updated properties.
We need to consider proxy settings change in order to generate a
new image.

The PR adds a check when downloading ISO for changed in proxy settings.
If there was a change, a new image will be generated and the new proxy
changes will be computed and saved for next ISO download check.

Signed-off-by: Moti Asayag <masayag@redhat.com>