Raz/mgmt 1245 (#172)
* MGMT-1245: added utils for checking pvc size

functions that will be used in the next commit, responsible for
extracting PVCs sizes from yaml or from existing environment and
comparing between PVCs and storing the largest size in the yaml.

* MGMT-1245: postgres deployment refactor - no logic change

1. split main to sub functions.
2. use log instead of print.
3. override values using yaml module.

* MGMT-1245: deploy postgres - check for existing pvc size

Use the logic of pvc_size_utils.py to override the storage size
of postgres PVC if the existing storage value is larger then the
default.

* MGMT-1245: scality deployment refactor - no logic change

1. split main to sub functions.
2. use log instead of print.
3. override values using yaml module.

* MGMT-1245: deploy scality - check for existing pvc size

Use the logic of pvc_size_utils.py to override the storage size
of scality PVC if the existing storage value is larger then the
default.