MGMT-1865: Add 'force' flag to job Deletion
If a cluster installation was failed due error in kube config
generation job, the assisted-installer didn't clear it on
reset command. The result is that user couldn't start a new
installation since the previous job was already exist.
Added 'force' flag that determines whether to delete the job
even if it failed, assume that if the user requested to reset
the cluster we no longer need that job.