Raz/mgmt 1644 (#187)
* MGMT-1644: Deployment: create directory per namesapce

First step to allow parallel execution of assisted-installer is
to store each namespace deployment files under a sub-directory
to avoid conflicts.

* MGMT-1644: Support multiple deployments in parallel

Add --server parameter to kubectl commands on local mode in order
to specify relevant minikube cluster to deploy, instead switching
between profiles. This is safer way to avoid conflicts in case of
parallel deployment.

* MGMT-1644: pvc_size_utils.py - small refactor

noticed that I assigned variable name that is identical to the
func name, changed that.

* MGMT-1644: Updated BUILD_FOLDER in Makefile