MGMT-1707: fix installing-pending-user-action event log
After host status is changed from installing to
installing-pending-user-action the agent is still tries to
register to the cluster.
Since there is no transition that handled registration from
source state installing-pending-user-action, the registration
always failed with error "Failed to register host: error
creating host metadata".
In this commit I extended the transition for register host
and return a specific error in such cases, which wont create
event log message on each try and prevent from the agent to
try to register again.