fixes three bugs in pkg/s3wrapper/filesystem.go
ListObjectsByPrefix was not accounting properly for the base directory, so it was
failing to identify the desired files.

ListObjectsByPrefix returned a slice of full paths on the local disk instead of
object references.

UploadStream used a 0-byte Read to identify the end of a file instead of
io.EOF. I don't fully understand why, but in practice, I observed 0-byte reads
happening without an EOF.

Fixing these three issues was required to get the subsystem test
"[only_k8s]Download cluster logs" to pass.

Thanks are due to @rwsu who also helped identify and resolve these bugs.