NO-JIRA improves commit message checking
Instead of assuming "origin" is the remote that corresponds to
github.com/openshift/assisted-service, the script now finds the remote
corresponding to that location. This works better for devs who use "origin" as
the name for their fork's remote.

If the commit is rejected, a message is shown citing the location of the file
where the commit message is preserved. Previously, it looked to the dev like
their commit message had just been thrown away.

In the interest of ensuring the dev reads the error message in case of commit
rejection, the script no longer prints the entire commit message before
displaying the helpful error message. By not throwing a wall of text at the dev,
they'll be more likely to see the error info itself.