OCPBUGSM-18040 Cluster installation fails with error : the container … (#407)
* OCPBUGSM-18040 Cluster installation fails with error : the container name \"assisted-installer\" is already in use.
Failure is caused by executing multiple install commands on the host.
Since only one container with the name "assisted-installer" can be executed just one install command is actually running and the rest return with error setting the host installation to failure
The trigger to this behavior is the time it take to pull the installer image from the downstream registry.
This patch mitigate the error by not moving the host to error in case the install command fail with that specific error code and stderr output (the container name \"assisted-installer\" is already in use) .

* OCPBUGSM-18040 Added subsystem test for installation step failure

* NO-ISSUE Revert "MGMT-1734 Add Jenkins stage to run assisted-test-infra (#268)"

This reverts commit b58eeb97d6a87ed6a6d670928c99faf812acaac4.