MGMT-1802 ports ignition generation from python Job to local golang (#185)
When run in-cluster or directly with podman, assisted-service will use
openshift-install locally to generate ignition files and make any modifications
(such as adding the status annotation to BareMetalHosts). For now,
openshift-install is embedded in the assister-service image, just as it has
been embedded in quay.io/ocpmetal/assisted-ignition-generator. Soon that will
be replaced with logic that extracts openshift-installer from a release.

The k8s Job that previously ran quay.io/ocpmetal/assisted-ignition-generator
has been removed.