MGMT-2320: swagger changes + autogenerated
In order to allow soft-deleted host and cluster models received deleted_at
flag. Each deletion will automatically be a soft-deletion, while the model
is still exists in database to be used when attempting to get logs.

In order to know which clusters were deleted but still exists in database,
new flag get_unregistered_clusters will determine whether to return a list
of unregistered clusters only.