MGMT-1780 Allow user to provide manifests for customizing cluster installation (#215)
When cluster is being installed using openshift-install, the user can
add manifests to be included in the generated bootstrap.ign and to
determine how the cluster should be configured prior to its installation.

The steps for generating the ignition files are as follow:
- Create install-config.yaml (handled by the service)
- Create manifests
- Add customization manifests
- Generate Ignition files

With the introduced manifests API, the user should be able to:
- Upload manifests to the cluster
- View manifests of the cluster
- Delete a manifest from the cluster
- Download a manifest

The assisted-service will extend the ignition generating to consider
cluster manifests if provided by the user.
If the user provides manifests, they will be stored on S3 prior to cluster
installation. When the cluster is being installed, the manifests will
be retrieved from S3 and placed under the '$INSTALL_DIR/manifests',
while the expect sub-folders should be 'openshift' and 'manifests' as
expected by the openshift-installer.

Signed-off-by: Moti Asayag <masayag@redhat.com>