MGMT-2411: swagger error models update + autogenerated
The fields of error/infra_error models are generated as pointers,
although we are always initializing all of the values when creating them.
Since the values are pointers the format of the generated error response
looks like this:
downloadClusterFilesConflict  &{Code:0xc00038aa70 Href:0xc00038aa80 ID:0xc000541e04 Kind:0xc00038aa90 Reason:0xc00038aaa0}

The most simplest way to fix this is to change the fields to be values
instead of pointers.