MGMT-2295 Give the test cluster a unique name (#597)
Previously all these clusters deployed to the various cloud
environments were a very generic "test-cluster" because we were
not providing anything different.

After this change the cluster will have a slightly more unique base
name but will also share a name with the openshift job that kicked
off the install which might be helpful in the future.

Fixes MGMT-2295