mgmt-2320 expose logs deleted clusters (#525)
* MGMT-2320: Added job to permanently delete old clusters and hosts

A job that will run once in an hour (configurable), delete all unscoped clusters
and hosts older then the duration defined in env. When deleting the cluster clean
all S3 files related to it and the cluster events.

* MGMT-2320: Updated ListClusters route

Will return also soft-deleted clusters if get_unregistered_cluster is set.

* MGMT-2320: Get cluster - return unscoped clusters

* MGMT-2320: Added job to permanently delete old clusters and hosts

A job that will run once in an hour (configurable), delete all unscoped clusters
and hosts older then the duration defined in env. When deleting the cluster clean
all S3 files related to it and the cluster events.

* MGMT-2320: getLogFileForDownload support deleted clusters logs

If cluster was unregistered but still existing in database we are able to return
its logs from S3. When attempting to get cluster logs try to search also for soft
deleted clusters. Since the cluster is not really deleted on deregister the events
should not be deleted yet.

* MGMT-2320: Clusters and hosts deletion will be executed only from leader