MGMT-2221 Host-specific ignition files (#466)
* MGMT-2221 Generate and upload a separate ignition for each host

This allows us to create each ignition with the /etc/hostname
content we want rather than making this change in the agent.

This will also make applying user customization to the host ignition
much easier.

* MGMT-2221 Create new endpoints for access to host-specific ignition files

One endpoint is primarily for the agent to use (../ignition/file).
This one will download the file. The other (../ignition)  will return
the file contents in the response.

* MGMT-2221 Move some utility functions out of the cluster api

These don't actually use any data from the manager and can just
be exported package functions.

This will make it much easier when testing as we can test the actual
functionality rather than mocking something that doesn't need to be
mocked.

* MGMT-2221 Extract and export the host ignition file name logic

* MGMT-2221 Add get and download host ignition API backend functions

* MGMT-2221 Update dummy ignition for subsystem tests

* MGMT-2221 Write out ignition files in parallel