MGMT-2857: Fix image expiration
Before the ISO generation optimization, we added a tag with the current
timestamp to the generated ISO and overwrote it if we wanted to reuse
the ISO.  The image expiration looked only at the tag to determine if it
should delete the image.
However, placing the tag on the initial image was accidentally left out
when the optimization code went in.  Thus, images that were never reused
do not have the timestamp tag, and are therefore never deleted.
This could be fixed by adding the tag back to the image generation flow.
However this would mean that we would need to manually clean old images
from deployments that do not have a tag.  Therefore the approach of this
commit is to check if the tag exists, and if not, fall back to the
object's creation time.