MGMT-2889 Allow registration of host in error state
It can happen that user rebooted the host manually after installation failure without changes in the cluster
if the registration return conflict or error then we have infinite number of events.
if the registration is blocked (403) it will break auto-reset feature.
So the best option is just accept the registration without changes in the DB