OCPBUGSM-20532: Allow register deleted hosts (#685)
* OCPBUGSM-20532: Allow register deleted hosts

If host is soft deleted by user, there is a chance that
it will try to register itself to the cluster again. In
such cases we should just remove the previews record of
the deleted host while registering it.

* OCPBUGSM-20532: Register host will be in transaction

In case of error while registering the host we don't want
to delete the soft deleted record of the previews host.
Therefore the host registration will be in a db transaction.