MGMT-2765 Run a subet of subsystem tests for onprem in Jenkins (#692)
* MGMT-2765 Run a subet of subsystem tests for onprem in Jenkins

We have a shortage of resources in Jenkins and this prevents
us from running all subsystem tests for onprem in parallel.
To ensure we have some test coverage, we decided to run a small
subset of the subsystem tests that can exercise the onprem code
paths.

The tests that constitute the small subset have been tagged with
[minimal-set] in the test definitions.

* MGMT-2765 set FOCUS and update clear-all to include clean-onprem