MGMT-2982 Fix image build race condition (#702)
Before this change it was possible for an image built for a PR
to end up deployed to the integration environment.

To fix, we should have built the initial image as the git sha
then only if we needed to push, tag latest.

The race happened like this:
1. build for master starts
2. build for pr starts
3. build for master finishes and tags latest
4. build for pr finishes and tags latest
5. push latest

Fixes MGMT-2982