KNIECO-2810: Add a new Dockerfile to enable ocp-ci
The existing Dockerfile for assisted-iso-create is expecting
a Docker build-arg, yet ci-operator does not support passing a
build argument.

A workaround is switching from using an ARG to an ENV, as ENVs
are supported in ocp-ci.

Since the work on this is still WIP, a new Dockerfile was created
in order to avoid any issues with the current functional Jenkins
pipeline.