MGMT-2943 Expose discovery ignition separately from iso (#691)
This commit uploads the discovery ignition to S3 in addition to
the iso with the ignition embedded. It also allows for the ignition
to be downloaded through the same file download endpoint we use for
other cluster files.

This will be used for cases where the rest of the iso will be
made available in some other manner.

Fixes MGMT-2943