MGMT-3103: reduce image expiration time and interval for local e2e tests (#752)
When deploying AI on minikube with local config, an iso is stored on s3
before each test. This cause the size of storage set to fill up and to
pods are starting to crash.
To fix this I reduce the IMAGE_EXPIRATION_TIME and
IMAGE_EXPIRATION_INTERVAL to 5m when running e2e tests on minikube.
Just need to add E2E_TESTS_MODE=1 when deploying installer configmap
and it will set the right config.
This solved the issue.