Add a migration generator and documentation about migrations (#596)
* NO-ISSUE Add some helper functions for migration tests

These will make it easier for developers to run just the migrations
that they need to to test new migrations

* NO-ISSUE Add a migration generation make target

This can be used by developers to generate a properly time stamped
migration file and test file which will contain a scaffolding for
how to write a new migration.

Immediately after generating the migration tests should run but will
fail as the new migration would not have been added to the list
in `migrations.all()`.

* NO-ISSUE Add migration docs

* NO-ISSUE Regenerate docs site