MGMT-3319: Integrating cluster list with OCM.
This will allow AMS to maintain assisted-installer's clusters in the
same list it does for OCP and OSD clusters.

We are creating a subscription object in AMS on cluster creation, later
on, when a cluster finishes its installation and the openshift_cluster_id
is "stable", then we will patch the subscription created for the cluster
with this ID.

This feature is turned on/off.

Signed-off-by: Yoni Bettan <ybettan@redhat.com>