MGMT-2758 Remove image from s3 when editing the discovery ignition (#765)
* MGMT-2758 Remove image from s3 when editing the discovery ignition

This will force the user to re-generate the image after making changes

Fixes MGMT-2758

* MGMT-2758 Remove the extra S3 call to check if an old iso exists

This adds an additional return value to the DeleteObject function
which tells the caller if an object was actually deleted or if one
was not found. Previously in the "not found" case, a success was
returned.