MGMT-3295: On installation delete old cluster logs
We saw in some triaging tickets that when we retrieve failed cluster
logs we can get hosts that were part of the previous installation.
In addition, if a host was re-registered with the same dns, we will
get duplicated log files after the second installation.
To avoid this we can simply delete previous installation log files
just before starting the new installation.