MGMT-3615 Set machine network CIDR as a pointer in cluster CRD (#920)
Machine network cidr should not be set to empty string by default, in some cases the service will set it the value, in those cases we will have a conflict so a way to prevent it is to set it as a pointer.
Pointers are used in the API in order to distinguish between between not set or set to empty string.