MGMT-3510 Allow static ip configuration in minimal iso initrd
This commit uses the same script and data we would typically
use in the discovery ignition to configure a static ip before
we try to download the rootfs on first boot.

The script then doesn't need to be provided to the ignition because
the coreos boot process copies the network files from /run/NetworkManager
to /etc/NetworkManager on the rootfs.

ref:
https://github.com/coreos/fedora-coreos-config/blob/076ba7c0917db6d2d2d12c82f0251d5eff32651a/overlay.d/05core/usr/lib/dracut/modules.d/35coreos-ignition/coreos-teardown-initramfs.sh#L69-L104