MGMT-3510 Alter the static ip config script to work during boot (#958)
For the minimal iso + static IP case we need to ensure we have
networking before we attempt to fetch the rootfs. This means that
we need to configure a static IP address early in the boot process.

In this environment we have a limited set of tools so to make the
script work for both scenarios this commit removes `find`, `envsubst`,
and `systemd-cat` from the script.

Additionally it adds a conditional to change the location to write
out the networking config. In the initrd environment, we write to
/run/NetworkManager, but in the booted system, we would write to
/etc/NetworkManager.