MGMT-3970 Make the iso editor scratch space base dir configurable
Before this change we would unconditionally extract isos to a
random directory in /tmp and would create new iso files directly
in /tmp

Now that base directory is configurable with the WORKSPACE_BASE_DIR
environment variable. The idea is that we may want this to be a PV
so we are not using the container filesystem.