OCPBUGSM-25012 Fixing bad bandwidth subsystem test (#1107)
The 'First success bad bandwidth' host subsystem test used
common.TestImageStatusesSuccess and changed its downloadRate for that
test. Afterwards, other tests using the same global variable, had a
malformed downloadRate.
The flakiness was caused because the tests are running in a random order.

In order to fix that bug, a new ContainerImageAvailability object was
created with a bad download rate.