MGMT-4167 Support both 3.1 and 3.2 ignition versions
This allows us to consume either version 3.1 or 3.2 ignition files
from openshift-installer without needing to know the OCP-to-ignition
version mappings before hand.

We do this by detecting the "real" version when initially parsing
the ignition then resetting the version string internally before
returning the version 3.2 config struct in `ignition.ParseTo32`.
This allows all of our other code to manipulate the ignition structure
using a single version rather than having to duplicate the logic
for each ignition version or extract the ignition version handling.

I intentionally did not use the higher level `config.Parse` function
here because it doesn't give us as easy access to the "real" version
in the byte string we pass in plus this would mean using ignition 3.3
which is marked as experimental and is probably subject to change.

This strategy leaves us open to potential bugs if we somehow start
using features from later versions of ignition either in our source
or from user customizations, but this solution is the best we can do
in the short term to support installing OpenShift 4.6, 4.7, and 4.8
in the same codebase.