MGMT-3370: Added error logs to RegisterClusterInternal (#1113)
Some validation failures are not printed to log on cluster
registration, but just returned as part of the api response.
Since we use the internal registration logic in the cluster
controller we can miss some errors that are not logged in
UpdateClusterInternal.
In this PR I updated the error logging in case of cluster
registration failure and made sure the error will be printed.