MGMT-4090 Use nmconnection settings relevant to current host
The discovery ignition file will contain nmconnection files for all of the nodes.
This patch is responsible to copy only the relevant files for to the current host
to the network-manager config directory.
Therefore the service to perform that action will run only once and before network
manager service.
The destination folder on the host is: `./NetworkManager/system-connections/static_connection.nmconnection`

Signed-off-by: Moti Asayag <masayag@redhat.com>