MGMT-4191 Clean up ignition package naming (#1124)
This will allow us to minimize changes when updating to newer ignition
versions in the future.

The strategy will always be to parse up to the latest supported
version then manipulate types that will not change in future minor
ignition versions (File, Config, Unit, etc).

Ideally the only place that will require a change will be where we
are specifically dealing with older versions.