MGMT-3970 Make the iso editor scratch space base dir configurable (#1063)
* MGMT-3970 Make the iso editor scratch space base dir configurable

Before this change we would unconditionally extract isos to a
random directory in /tmp and would create new iso files directly
in /tmp

Now that base directory is configurable with the WORKSPACE_BASE_DIR
environment variable. The idea is that we may want this to be a PV
so we are not using the container filesystem.

* MGMT-3970 Make env var names iso-specific

* MGMT-3970 Update go-diskfs

Pulls in https://github.com/diskfs/go-diskfs/pull/101 which allows
us to specify the location of the workspace directory used to build
the iso

* MGMT-3970 Pass the existing iso directory to go-diskfs

This both removes the extra copy of the files created when we're adding
them to go-diskfs' staging area as well as allows go-diskfs to remove
the directory for us after we're done creating the iso.