KNIDEPLOY-4151 Operator deploys the assisted installer  (#1073)
* KNIDEPLOY-4151 Create manifests without deploying them to minikube or OCP cluster

Added --apply-manifest argument to deploy_ python scripts.
Setting it to False generates the resource manifests without
applying them.

This option is used by the create-ocp-manifests Makefile target.
The operator bundle will use this target to generate the ConfigMap,
Deployments, etc.. and include them into the CSV and bundle manifests.

* KNIDEPLOY-4151 Adds bundle creation

* KNIDEPLOY-4151 Remove name prefix and add public-s3-secrets

* KNIDEPLOY-4151 Create route for service and update SERVICE_BASE_URL

* KNIDEPLOY-4151 Add kustomizations for DEPLOY_TARGET, CREATE_SERVICE_ROUTE

* KNIDEPLOY-4151 Add http:// to route url

* KNIDEPLOY-4151 Cleanup code and add missing bundle.Dockerfile

* KNIDEPLOY-4151 Update ServiceBaseURL in job and instruction configs

* KNIDEPLOY-4151 Add additional rbac roles to bundle

* KNIDEPLOY-4151 Update assisted-service image through Kustomize

* KNIDEPLOY-4151 Operator bundle build and deployment doc

* KNIDEPLOY-4151 Add operator-sdk to Dockerfile.assisted-service-build

Also removed kustomize target because it is already in Dockerfile.assisted-service-build.

Renamed VERSION to OPERATOR_VERSION to avoid conflict.

* KNIDEPLOY-4151 Restore image to quay.io/ocpmetal and cleanup comments

* KNIDEPLOY-4151 Fix image patch

* KNIDEPLOY-4151 Document useful kustomize operator options/patches

* KNIDEPLOY-4151 Garbage collect route by adding ownerReference to service

Add ownerReference to service to ensure the route is garbage
collected when the operator is deleted. If not garbage collected,
it prevents the operator from being reinstalled.

* KNIDEPLOY-4151 Remove intermediate resource yaml files

They are generated by bundle process and are removed because there
isn't a need to archive them into git. Also removed config/certmanager
and config/samples as they aren't in use.

Renamed bundle to operator-bundle and bundle-build to
operator-bundle-build.

* KNIDEPLOY-4151 Add volume mount for operator deployed postgres

Volume mount was disabled for OCP deployments.

Added additional flag --persistent-storage to indicate if persistent
storage is desired. Decoupled choice of persistent/ephemeral storage
from deploy target.

* KNIDEPLOY-4151 Remove ui from operator deployment and fix repo in PROJECT

* KNIDEPLOY-4151 Create route using init container

* KNIDEPLOY-4250 Update CRDs defined in CSV bases

* MGMT-4172 Init container to create random password and secret for
postgres

Replaces the postgres secret yaml embedded in the operator.

* MGMT-4172 Add ui back for testing/debugging

* MGMT-4172 Remove bundle.Dockerfile

It is auto-generated by "operator-sdk generate bundle".

* MGMT-4172 Prefixed roles with assisted-service

* MGMT-4172 More description on postgres PVC requirement

* MGMT-4172 Ensure PERSISTENT_STORAGE is set correctly, cleanup code