MGMT-4184: Use the full version during the deployment
By using the full version as key for the available versions in the
OPENSHIFT_VERSIONS input, we' will instruct the assited-installer to use
that version, instead of the short form. This will allow to have more
detailed information in the database (OpenshiftVersion will be set to
the full version in the DB) as well as more granular metrics in
the monitoring tools being used.

We could keep the current json files and data as they are but, by being
explicit, we will also instruct on what we thing is the recommended
thing to do.