MGMT-4339 Add indication for discovery image generation (#1206)
In order to allow the discovery image to be re-generated, an indication
is added to describe the state of it.
When the discovery image is generated, the indicator will be set to false
and only after successful image generation it will be set to true.

The indicator will prevent the case in which a failed attempt to generate
an image will result with storing the image create parameters on the
cluster. In that case, the next attempt to generate the ISO will falsely
present as the image hasn't been changed without considering the new
parameters of the cluster.

Signed-off-by: Moti Asayag <masayag@redhat.com>