MGMT-4274 Allow for additional authenticator types (#1201)
* MGMT-4274 Rename auth_handler_utils to jwk

* MGMT-4274 Create an Authenticator interface

The underlying type of Authenticator we get will depend on the new
AUTH_TYPE environment variable which replaces ENABLE_AUTH

Right now we only have two types (rhsso and none) which gives us
parity with the existing behavior, but this will allow an easier
addition of a local type which will only be used of on-prem agent
authentication.

https://issues.redhat.com/browse/MGMT-4274

* MGMT-4274 Use the new authenticator interface

Also replace EnableAuth usages with appropriate AuthType ones.

https://issues.redhat.com/browse/MGMT-4274