MGMT-4428 Add publish target and set tag for origin-cli to latest
Adds the assisted-service-operator-bundle to the existing publish
target so that a new operator bundle image is published to quay.io
when PRs are merged.

origin-cli tag is set to latest matching Dockerfile.assisted-service.

Removes conditional on log_image_revision. The logging will now warn
that the image revision cannot be inspected.

Default bundle parent directory to BUILD_DIR.