MGMT-4275 Add local JWT authentication (#1228)
* MGMT-4275 Add local JWT authentication

This uses ecdsa keys to keep the resulting token signing string small.
This will make it a better fit for using in urls later.

* MGMT-4275 Validate the token type in the parser

Let the library ensure the algorithm specified by the token
header is what we expect and add tests for cases where the
algorithm is incorrect.

* MGMT-4275 Validate the cluster_id token claim

The only requirement we have here is that the cluster id
must exist in the database. We are doing this to implement a
rudimentary token revocation strategy.

* MGMT-4275 Add type assertions for the authenticators