MGMT-4428 Create Dockerfile to build bundle image in prow (#1230)
* MGMT-4428 Create Dockerfile to build bundle image in prow

* MGMT-4428 Upgrade oc client from 4.6.0 to 4.7.1

* MGMT-4428 Fetch oc client using COPY --from=quay.io/openshift/origin-cli

instead of curl. Using COPY is more reliable than curl because
it can use a locally stored version of the image if one is already
present.

* MGMT-4428 Add publish target and set tag for origin-cli to latest

Adds the assisted-service-operator-bundle to the existing publish
target so that a new operator bundle image is published to quay.io
when PRs are merged.

origin-cli tag is set to latest matching Dockerfile.assisted-service.

Removes conditional on log_image_revision. The logging will now warn
that the image revision cannot be inspected.

Default bundle parent directory to BUILD_DIR.