MGMT-4513: Don't do anything if we update host with unchanged info.
The controller keeps updating host installation progress until all hosts
are installed or they are all in error.

In case the user is resetting the cluster, there is a time window in
which the controller still runs and continues to update the hosts,
therefore events are spammed until k8s kills the controller once 2
masters aren't part of the cluster anymore.

Signed-off-by: Yoni Bettan <ybettan@redhat.com>