NO-ISSUE - attempt to logout from problematic registries before running CI
We've recently seen some issues with the registry.svc.ci.openshift.org registry. It seems that someone might have logged in to those registries in some job and it affected future CI jobs that came after it. This commit makes it so there's a more 'clean slate' before trying to run the CI