MGMT-4205 Cluster completes on monitored operators status (#1276)
As part of the cluster monitoring - when the cluster is in finalizing
state - it would check if the cluster had completed its installation as well.
The completion checks whether kubeconfig was uploaded to S3 server and
as well all the cluster monitored operators are available.

In order to be backward-compatible, we still keep the
CompleteInstallation API endpoint.
The new monitoring feature is available only in case a cluster has a
positive number of monitored operators.

Kubeconfig file would be able to be downloaded when the cluster is either
in Finalizing state or Installed state.

Updated subsystem tests completeInstallation function to be aligned with
the new process - updating cluster ingress certificate and reporting monitored
operators as available.