MGMT-4425 Add presigned url authentication (#1347)
* MGMT-4425 Add query based token auth

This will always fail for rhsso, always succeed for none,
and validates the same way as agent auth for local.

* MGMT-4425 Use the go-openapi function for fetching the token

This will handle getting the token from either the url
or the header. Before we were assuming the token would
always be in the header which isn't the case anymore.

* MGMT-4425 Move token and key generation to a separate package

This will make it easier to share generic token and key
generation functions across tests and application code.

Before, we had a function for generating a token and key
pair in pkg/auth and a test doing pretty much the same thing
in internal/cluster

* MGMT-4425 Add a helper function to sign urls

The caller will be responsible for creating the initial
valid URL. This new function will add a signed token
for the given cluster id.

* MGMT-4425 Sign image download url when using local auth