MGMT-4356: Deploy only single openshift version
This PR add a switch to tools/handle_ocp_versions.py script, that can
decide to deploy only single OCP version. If user pass the env var
SINGLE_VERSION_ONLY=1 it will first take --ocp-override version image,
and use it, if not set it will use OPENSHIFT_VERSION env var, if not set
it will take the latest version in the OPENSHIFT_VERSIONS definition.

This patch is needed for our CI so we can improve deployment time and
don't download RHCOS ISO images which we don't want to test anyway.

Signed-off-by: Ondra Machacek <omachace@redhat.com>