NO-ISSUE Fix boot file upload race
Previously on the first deploy of the service with multiple
replicas each pod would set `haveLatestMinimalTemplate` to true
and then block on the leader lock. When a pod got the lock it would
force an upload of the minimal template file even if it had been
uploaded by another replica already.

Additionally, we need to keep the check for the template outside
of the version loop as the template version marker file is shared
between all iso versions.

As a solution we take the lock once and only check if we have the
latest minimal template while holding the lock. Then we can safely
upload the boot files for all versions in a single replica. By the
time the other replicas get the lock they will see that the template
version is up to date and will no-op the uploads.

This issue was causing problems in ci tests as the api for the first
replica was marked ready while the other replicas were still uploading
minimal isos. It looks likely that this situation lead to a failed
ci job where the minimal iso couldn't be downloaded from scality.